# api-gateway/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "JobsController_create",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - requires RECRUITER or ADMIN role"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new job posting",
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "operationId": "JobsController_findAll",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in title and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "OPEN",
                "CLOSED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "employmentType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "FULL_TIME",
                "PART_TIME",
                "CONTRACT",
                "INTERNSHIP"
              ]
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "description": "Filter by department",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salaryMin",
            "required": false,
            "in": "query",
            "description": "Filter by minimum salary",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "salaryMax",
            "required": false,
            "in": "query",
            "description": "Filter by maximum salary",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "skills",
            "required": false,
            "in": "query",
            "description": "Filter by required skills (comma-separated)",
            "schema": {
              "example": "NestJS,TypeScript,PostgreSQL",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "title",
                "salaryMin"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated jobs list"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all jobs with pagination and filters",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "operationId": "JobsController_findOne",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID (UUID)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a job by ID",
        "tags": [
          "Jobs"
        ]
      },
      "put": {
        "operationId": "JobsController_update",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID (UUID)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not the owner or admin"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a job posting",
        "tags": [
          "Jobs"
        ]
      },
      "delete": {
        "operationId": "JobsController_remove",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID (UUID)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not the owner or admin"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a job (soft delete)",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "TalentFlow AI API",
    "description": "API documentation for the TalentFlow AI API Gateway",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your access token"
      }
    },
    "schemas": {
...
      "JobRequirementsDto": {
        "type": "object",
        "properties": {
          "skills": {
            "example": [
              "React",
              "Node.js"
            ],
            "description": "Required skills for the job",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "experience": {
            "type": "string",
            "example": "3+ years",
            "description": "Minimum experience expectation"
          }
        }
      },
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Senior Full Stack Developer",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "example": "We are looking for an experienced developer..."
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "location": {
            "type": "string",
            "example": "Remote"
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "CONTRACT",
              "INTERNSHIP"
            ],
            "default": "FULL_TIME",
            "example": "FULL_TIME"
          },
          "salaryMin": {
            "type": "number",
            "example": 80000,
            "minimum": 0
          },
          "salaryMax": {
            "type": "number",
            "example": 120000,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "OPEN",
              "CLOSED",
              "ARCHIVED"
            ],
            "default": "DRAFT",
            "example": "DRAFT"
          },
          "requirements": {
            "example": {
              "skills": [
                "React",
                "Node.js"
              ],
              "experience": "3+ years"
            },
            "description": "Structured job requirements",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobRequirementsDto"
              }
            ]
          }
        },
        "required": [
          "title",
          "employmentType",
          "status"
        ]
      },
      "JobResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "department": {
            "type": "object",
            "nullable": true
          },
          "location": {
            "type": "object",
            "nullable": true
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "CONTRACT",
              "INTERNSHIP"
            ]
          },
          "salaryMin": {
            "type": "object",
            "nullable": true
          },
          "salaryMax": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "OPEN",
              "CLOSED",
              "ARCHIVED"
            ]
          },
          "requirements": {
            "type": "object",
            "nullable": true
          },
          "createdById": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "department",
          "location",
          "employmentType",
          "salaryMin",
          "salaryMax",
          "status",
          "requirements",
          "createdById",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Senior Full Stack Developer",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "example": "We are looking for an experienced developer..."
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "location": {
            "type": "string",
            "example": "Remote"
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "CONTRACT",
              "INTERNSHIP"
            ],
            "default": "FULL_TIME",
            "example": "FULL_TIME"
          },
          "salaryMin": {
            "type": "number",
            "example": 80000,
            "minimum": 0
          },
          "salaryMax": {
            "type": "number",
            "example": 120000,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "OPEN",
              "CLOSED",
              "ARCHIVED"
            ],
            "default": "DRAFT",
            "example": "DRAFT"
          },
          "requirements": {
            "example": {
              "skills": [
                "React",
                "Node.js"
              ],
              "experience": "3+ years"
            },
            "description": "Structured job requirements",
            "allOf": [
              {
                "$ref": "#/components/schemas/JobRequirementsDto"
              }
            ]
          }
        }
      },
...
    }
  }
}